Added VJB and JDB headshots and code refactoring for modules.
